+++ gatsby-config.js
@@ -34,8 +34,8 @@ module.exports = {
         accessToken: process.env.GATSBY_CONTENTFUL_ACCESS_TOKEN,
       },
     },
-    // "gatsby-plugin-image",
-    // "gatsby-plugin-sharp",
-    // "gatsby-transformer-sharp",
+    "gatsby-plugin-image",
+    "gatsby-plugin-sharp",
+    "gatsby-transformer-sharp",
   ],
 };
