+++ src/pages/login.tsx
@@ -3,7 +3,7 @@ import { navigate } from "gatsby";
 import { Helmet } from "react-helmet";
 import qs from "query-string";
 
-import { UserContext } from "../context/UserContext";
+import UserProvider, { UserContext } from "../context/UserContext";
 import Layout from "../components/templates/Layout";
 
 const LoginPage = ({ location }: any) => {
@@ -17,19 +17,27 @@ const LoginPage = ({ location }: any) => {
     }, 1000);
   });
 
+  const handleClick = () => {
+    setUser("Clicked");
+  };
+
   return (
     <Layout>
-      <Helmet>
-        <title>Authenticating...</title>
-      </Helmet>
+      <UserProvider>
+        <Helmet>
+          <title>Authenticating...</title>
+        </Helmet>
 
-      <div className="container max-w-5xl mx-auto">
-        <header>
-          <h2 className="text-4xl font-bold text-center text-blue-400">
-            Authenticating your account...
-          </h2>
-        </header>
-      </div>
+        <div className="container max-w-5xl mx-auto">
+          <header>
+            <h2 className="text-4xl font-bold text-center text-blue-400">
+              Authenticating your account...
+            </h2>
+          </header>
+        </div>
+
+        <button onClick={handleClick}>Clicked</button>
+      </UserProvider>
     </Layout>
   );
 };
